new_release New_add_doc (#71)
* Avoid simple pass in Exceptions

* Restrict coverage folders in GH action

* Removed ipython from poetry

* Refactored copy_log in cli.py

* How to setup for dev

* Improved exception handling

* Added a janitor fixture for pytest

* Fixed ver_today with pcregrep (linux/macos)

* Remove ipython dev dependency

* Test verboless

* Fixed bug for sorted atomPairs

* Test ILDN with -z exception

* Test ILDN with -z exception improved

* Corrected logger().exception use

* Install poetry via curl

* Pytest in random order

* Test for corrupt glycam

* Added sphinx and docs

* Adding docs, just starting

* Fixed some badges

* Added info about documenting

* Improved doc example

* Trying to refine actions

* Use dynamic year for copyright

* Tests for mol and mdl formats

* Test sqm fails due to wrong charge input

* Typings and docs

* Test for ekFlag

* Try/exception around setResNameCheckCoords()

* Remove try/exception around os.chdir(mol.rootDir)

* Tests about resname

* tests for empty amber files

* Test mol geometry

* Check resname

* convertSmilesToMol2: redundant obabel check

* Increasing coverage about charges

* Fixed regex in ver_today.sh

* Poetry added pyupgrade, pre-commit and sphinx

* Make check_acpype.py to work, still need more

* Pyupgraded

* Added printErrorQuoted

* pre-commit end-of-file-fixer

* pid and docstrings errors cleared

+++ acpype/logger.py
@@ -44,6 +44,8 @@ class LogFormatter(logging.Formatter):
 
 
 def copy_log(molecule):
+    if not os.path.exists(molecule.absHomeDir):
+        raise UnboundLocalError
     local_log = os.path.join(molecule.absHomeDir, "acpype.log")
     if os.path.exists(local_log):
         os.remove(local_log)
